requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Neo4j
neo4j==5.16.0

# Caching & Queue
redis==5.0.1
celery==5.3.6
flower==2.0.1

# PDF Processing
pypdf==4.0.1
pdfplumber==0.10.4

# NLP & ML
torch==2.1.2
transformers==4.37.2
sentence-transformers==2.3.1
rank-bm25==0.2.2
spacy==3.7.2

# LLM Clients
openai==1.10.0
anthropic==0.18.0
tiktoken==0.5.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0

# Observability
structlog==24.1.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Data Science
numpy==1.26.3
scikit-learn==1.4.0
